Thread task_executor into BeaconChain
