+++ beacon_node/execution_layer/src/lib.rs
@@ -1204,16 +1204,15 @@ impl ExecutionLayer {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::test_utils::{
-        ExecutionLayerRuntime, MockExecutionLayer as GenericMockExecutionLayer,
-    };
+    use crate::test_utils::MockExecutionLayer as GenericMockExecutionLayer;
+    use task_executor::test_utils::TestRuntime;
     use types::MainnetEthSpec;
 
     type MockExecutionLayer = GenericMockExecutionLayer<MainnetEthSpec>;
 
     #[tokio::test]
     async fn produce_three_valid_pos_execution_blocks() {
-        let runtime = ExecutionLayerRuntime::default();
+        let runtime = TestRuntime::default();
         MockExecutionLayer::default_params(runtime.task_executor.clone())
             .move_to_terminal_block()
             .produce_valid_execution_payload_on_head()
@@ -1226,7 +1225,7 @@ mod test {
 
     #[tokio::test]
     async fn finds_valid_terminal_block_hash() {
-        let runtime = ExecutionLayerRuntime::default();
+        let runtime = TestRuntime::default();
         MockExecutionLayer::default_params(runtime.task_executor.clone())
             .move_to_block_prior_to_terminal_block()
             .with_terminal_block(|spec, el, _| async move {
@@ -1246,7 +1245,7 @@ mod test {
 
     #[tokio::test]
     async fn verifies_valid_terminal_block_hash() {
-        let runtime = ExecutionLayerRuntime::default();
+        let runtime = TestRuntime::default();
         MockExecutionLayer::default_params(runtime.task_executor.clone())
             .move_to_terminal_block()
             .with_terminal_block(|spec, el, terminal_block| async move {
@@ -1263,7 +1262,7 @@ mod test {
 
     #[tokio::test]
     async fn rejects_invalid_terminal_block_hash() {
-        let runtime = ExecutionLayerRuntime::default();
+        let runtime = TestRuntime::default();
         MockExecutionLayer::default_params(runtime.task_executor.clone())
             .move_to_terminal_block()
             .with_terminal_block(|spec, el, terminal_block| async move {
@@ -1282,7 +1281,7 @@ mod test {
 
     #[tokio::test]
     async fn rejects_unknown_terminal_block_hash() {
-        let runtime = ExecutionLayerRuntime::default();
+        let runtime = TestRuntime::default();
         MockExecutionLayer::default_params(runtime.task_executor.clone())
             .move_to_terminal_block()
             .with_terminal_block(|spec, el, _| async move {
