+++ common/task_executor/Cargo.toml
@@ -5,9 +5,10 @@ authors = ["Sigma Prime <contact@sigmaprime.io>"]
 edition = "2021"
 
 [dependencies]
-tokio = { version = "1.14.0", features = ["rt"] }
+tokio = { version = "1.14.0", features = ["rt", "rt-multi-thread"] }
 slog = "2.5.2"
 futures = "0.3.7"
 exit-future = "0.2.0"
 lazy_static = "1.4.0"
 lighthouse_metrics = { path = "../lighthouse_metrics" }
+sloggers = { version = "2.1.1", features = ["json"] }
