+++ common/task_executor/src/lib.rs
@@ -1,4 +1,5 @@
 mod metrics;
+pub mod test_utils;
 
 use futures::channel::mpsc::Sender;
 use futures::prelude::*;
